requirements.txt

flask
pandas
sentence-transformers
scikit-learn
numpy
PyPDF2
python-docx
spacy
